backend-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.12
httpx==0.28.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0
python-dotenv==1.0.0
elasticsearch==8.11.0
sentence-transformers>=2.3.0
huggingface_hub>=0.30.0,<1.0.0
torch>=1.13.0
transformers>=4.21.0
